+++ packages/test/src/test-integration-codec-old.ts
@@ -1,5 +1,5 @@
 import { RUN_INTEGRATION_TESTS, ByteSkewerPayloadCodec } from './helpers';
-import { runIntegrationTests } from './integration-tests';
+import { runIntegrationTests } from './integration-tests-old';
 
 if (RUN_INTEGRATION_TESTS) {
   runIntegrationTests(new ByteSkewerPayloadCodec());
