+++ packages/test/src/test-integration-old.ts
@@ -1,5 +1,5 @@
 import { RUN_INTEGRATION_TESTS } from './helpers';
-import { runIntegrationTests } from './integration-tests';
+import { runIntegrationTests } from './integration-tests-old';
 
 if (RUN_INTEGRATION_TESTS) {
   runIntegrationTests();
